% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funtions.R
\name{AupdatedB}
\alias{AupdatedB}
\title{replace parameters data}
\usage{
AupdatedB(DataA, DataB, ons)
}
\arguments{
\item{DataA}{DataA}

\item{DataB}{DataB}

\item{ons}{by on colume}
}
\value{
Pdata.frame
}
\description{
replace parameters data
}
