% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BPring}
\alias{BPring}
\title{Example data of BTRtools: observed value of tree ring}
\format{
A \code{data.frame} containing 5 columns :

\strong{Year} : Year,

\strong{MRW} : Mean ring width,

\strong{MaxLA} : Maxium lumen area in ring,

\strong{CD} : Vessel density,

\strong{RCTA} : Vessel fraction.
}
\source{
A process-based model of climate-driven xylogenesis and tree-ring formation in broad-leaved trees (BTR).\link{DOI:10.1093/treephys/tpae127}
}
\usage{
data(BPring)
}
\description{
Example data of BTRtools: observed value of tree ring
}
\keyword{datasets}
