% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/btmcmc.R
\name{btmcmc}
\alias{btmcmc}
\title{btmcmc}
\usage{
btmcmc(logfile)
}
\arguments{
\item{logfile}{The name of the logfile of the BayesTraits analysis.}
}
\value{
A data frame containing the sample from the BayesTrait mcmc.
}
\description{
Returns the full mcmc object from a BayesTraits log file. This
is used inside plot functions and so on, but might be useful for
other MCMC manipulations and so on.
}

