% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funtions.R
\name{BtrResRemake}
\alias{BtrResRemake}
\title{Plot data for BTRres}
\usage{
BtrResRemake(
  dtlist = dataResBTR$ResBTR,
  dtin = dataInBTR,
  years = input$ShowY,
  yC,
  yV
)
}
\arguments{
\item{dtlist}{data list of shiny}

\item{dtin}{data Input of shiny}

\item{years}{InputYears}

\item{Obsdtlist}{description}
}
\value{
Plot data
}
\description{
Plot data for BTRres
}
