% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btrtools_desc.R
\docType{package}
\name{BTRTools}
\alias{BTRTools}
\alias{BTRTools-package}
\title{A package that makes it easier to process and analyse the output of BayesTraits.}
\description{
A package that makes it easier to process and analyse the output of BayesTraits.
}
\author{
Henry Ferguson-Gow
}
