% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funtions.R
\name{C3mod}
\alias{C3mod}
\title{Calculate fuc}
\usage{
C3mod(method, a, b, c, age)
}
\arguments{
\item{method}{method}

\item{a}{As}

\item{b}{Bs}

\item{c}{Cs}

\item{age}{age}
}
\value{
res
}
\description{
Calculate fuc
}
