% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funtions.R
\name{CalculateCells}
\alias{CalculateCells}
\title{CalCells}
\usage{
CalculateCells(
  CBparam = dataInCell$CBParam,
  Cparam = dataInCell$VParam,
  gRLi = 1,
  DroughtIndex = c(0.5, 0.95)
)
}
\arguments{
\item{CBparam}{CBparam}

\item{Cparam}{Cparam}

\item{gRLi}{gRLi}

\item{DroughtIndex}{DroughtIndex}
}
\value{
Ress
}
\description{
CalCells
}
