% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funtions.R
\name{CellGrwothData}
\alias{CellGrwothData}
\title{For table 2,3 calculate cell growth}
\usage{
CellGrwothData(param, wgR, dry, fgRLi, vgRLi)
}
\arguments{
\item{param}{parameters}

\item{wgR}{wgR gR 0-1}

\item{dry}{dry factor 0-1}

\item{vgRLi}{L_i for vessel gR}

\item{fgRli}{L_i for fiber gR}
}
\value{
cell daily traits and new param
}
\description{
For table 2,3 calculate cell growth
}
