% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funtions.R
\name{CellLine}
\alias{CellLine}
\title{Cell Line simulate}
\usage{
CellLine(ObsV, ObsF)
}
\arguments{
\item{ObsV}{Obsdata vessel}

\item{ObsF}{Obsdata fiber}
}
\value{
Cell traits line
}
\description{
Cell Line simulate
}
