% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Cells}
\alias{Cells}
\title{Example data of BTRtools: observed value of fiber cells, output file of ROXAS}
\format{
A \code{data.frame} containing 33 columns :
}
\source{
A process-based model of climate-driven xylogenesis and tree-ring formation in broad-leaved trees (BTR).\link{DOI:10.1093/treephys/tpae127}
}
\usage{
data(Cells)
}
\description{
Example data of BTRtools: observed value of fiber cells, output file of ROXAS
}
\keyword{datasets}
