% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Clims}
\alias{Clims}
\title{Example data of BTRtools: Cliamte data}
\format{
A \code{data.frame} containing 11 columns :

\strong{Year} : Year,

\strong{DOY} : Day of years,

\strong{Month} : Month,

\strong{Day} : Day,

\strong{TEM} : Temperature,

\strong{gE} : Grwoth rate by light,

\strong{soilM} : Soil moisture,

\strong{Ls} : Day length,

\strong{rootd} : Root zone depth,

\strong{dL_i} : dealt day length,

\strong{VPD} : vapor Pressure Deficit.
}
\source{
A process-based model of climate-driven xylogenesis and tree-ring formation in broad-leaved trees (BTR).\link{DOI:10.1093/treephys/tpae127}
}
\usage{
data(Clims)
}
\description{
Example data of BTRtools: Cliamte data
}
\keyword{datasets}
