% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareStones.R
\name{compareStones}
\alias{compareStones}
\title{compareStones}
\usage{
compareStones(comp, simp)
}
\arguments{
\item{comp}{The name of the stepping stones log file of the more complex of the two BayesTraits models.}

\item{simp}{The name of the stepping stone log file of the more simple of the two BayesTraits models.}
}
\value{
The log Bayes factor of the comparison between the more complex and the more simple model.
}
\description{
A function to extract the marginal likelihood from the stepping
stone sampler output of two BayesTraits runs, and work out the
log Bayes Factors.
}
\examples{
compareStones("complex.model.stones.log.txt", "simple.model.stones.log.txt")
}
\keyword{Bayes}
\keyword{factors}
\keyword{hypothesis}
\keyword{significance}
\keyword{stepping}
\keyword{stone}
\keyword{testing}
