% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/compPosts.R
\name{compPosts}
\alias{compPosts}
\title{compPosts}
\usage{
compPosts(logs, pars)
}
\arguments{
\item{logs}{The name of the trait data file on which BayesTraits was run, or a vector of >1 names if comparing between >1 logs.}

\item{pars}{A vector containing the names of two parameters to be compared. Must be a single parameter if comparing between two logs. Can
be subsetted from the output of getParams().}
}
\description{
Compares histograms of one or more parameters from the same output file, or one parameter from one or more output files.
}
\examples{
plotPosterior(cool-data.txt, c("Lh", "Alpha 1"))
plotPosterior(cool-data.txt, params[c(1:2)])
}
\keyword{compare}
\keyword{distribution}
\keyword{histogram}
\keyword{plot}
\keyword{posterior}

