% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funtions.R
\name{compute_daylength}
\alias{compute_daylength}
\title{Compute_daylengthfactor}
\usage{
compute_daylength(phi)
}
\arguments{
\item{phi}{Fix parameters}
}
\value{
day length and standardization day length
}
\description{
Compute_daylengthfactor
}
