% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjpp.R
\name{createCounts}
\alias{createCounts}
\alias{createCountsTable}
\title{createCounts
Creates the counts table for the rjpp.}
\usage{
createCountsTable(extree, meanbl)
}
\arguments{
\item{extree}{The time tree}

\item{meanbl}{The output of meanBranches}
}
\description{
createCounts
Creates the counts table for the rjpp.
}
\keyword{internal}
