% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funtions.R
\name{DailCellGr}
\alias{DailCellGr}
\title{Daily cell growth}
\usage{
DailCellGr(cell = dailycells, Cparam, gRLi, i_dry, gREL = 0.65, DOY = 1)
}
\arguments{
\item{cell}{cell}

\item{Cparam}{Cparam}

\item{gRLi}{gRLi}

\item{i_dry}{i_dry}

\item{gREL}{gREL}

\item{DOY}{description}
}
\value{
Cells
}
\description{
Daily cell growth
}
