% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findShifts.R
\name{findRateShifts}
\alias{findRateShifts}
\title{findRateShifts}
\usage{
findRateShifts(PP, threshold, type)
}
\arguments{
\item{PP}{an output from rjpp}

\item{threshold}{the threshold over which scalars are considered interesting.}

\item{type}{Node or branch}
}
\description{
Return the node or branch numbers with rate scalars applied over the threshold number.
}
\keyword{internal}
