% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findShifts.R
\name{findShifts}
\alias{findShifts}
\title{findShifts}
\usage{
findShifts(PP, scalar, threshold)
}
\arguments{
\item{PP}{The psotprocessor (localscalrPP) output.}

\item{scalar}{The scalar to find and plot from the post processor - delta/lambda/kappa/node/branch. If rate then the nodes and branches that are given scalars over the threshold are returned.}

\item{threshold}{Threshold of probability in posterior to display deltas for, defaults to zero (i.e. shows all deltas shaded proportionally to the posterior probability)}
}
\description{
Returns the node numbers of nodes that have been scaled in the posterior over the specified threshold. Works for rate scalars, and transformations. Also can return branch scalars, if rates. Only looks at the placement of scalars, not the cumulative downstream effects of scalars.
}
