% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findShifts.R
\name{findTransShifts}
\alias{findTransShifts}
\title{findTransShifts}
\usage{
findTransShifts(PP, threshold, cl)
}
\arguments{
\item{PP}{an output from rjpp}

\item{threshold}{The threshold over which transformations are considered "significant".}
}
\description{
return the node numbers that have had transformations applied over a given threshold.
}
\keyword{internal}
