% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeStructFuncs.R
\name{getMRCAbtr}
\alias{getMRCAbtr}
\title{getMRCAbtr}
\usage{
getMRCAbtr(x, tree, rjtaxa)
}
\arguments{
\item{x}{A vector of taxa names}

\item{tree}{A phylogeny of class "phylo" (generally the time tree used as input to BayesTraits)}

\item{rjtaxa}{The taxa translations as output from BayesTraits}
}
\description{
This an extension of apes's getMRCA that enables the return of a tip, or an MRCA. Translates taxa codes (BayesTraits) to proper tip labels. Useful only in post-processing.
}
\keyword{internal}
