% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParams.R
\name{getParams}
\alias{getParams}
\title{getParams}
\usage{
getParams(logfile)
}
\arguments{
\item{logfile}{The name of the trait data file on which BayesTraits was run.}
}
\value{
A vector of the parameter names of the model logfile results from.
}
\description{
This functions returns the names of the estimated parameters from a BayesTraits
analysis logfile for input into plotting and other analysis functions.
}
\examples{
getParams("cool-data.txt")
}
\keyword{parameters}
