% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/getPostMean.R
\name{getPostMean}
\alias{getPostMean}
\title{getPostMean}
\usage{
getPostMean(logfile, params)
}
\arguments{
\item{logfile}{The name of the logfile from a BayesTraits analysis.}

\item{params}{A vector of one or more parameters of interests (e.g. a subset of getParams output)}
}
\value{
A data frame of 3 columns, the parameter name, the mean of the parameter and its standard deviation.
}
\description{
A function to retuen the mean and SD of the posterior of one or more
parameters from a BayesTraits analysis.
}
\examples{
params <- getParams("cool-data.log.txt")
getPostMean("cool-data.log.txt", params[c(2:5)]
getPostMean("cool-data.log.txt", "Lh")
}
\keyword{deviation}
\keyword{mean}
\keyword{posterior}
\keyword{standard}

