% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ggAutoCor.R
\name{ggAutoCorr}
\alias{ggAutoCorr}
\title{ggAutoCorr}
\usage{
ggAutoCorr(pars, conf, max.lag = NULL, min.lag = 0)
}
\arguments{
\item{pars}{The paramater you want to see an autocorrelation plot for.}

\item{conf}{The confidence level you want to see the limits for.}

\item{max.lag}{Maximum lag}

\item{min.lag}{Minimum lag}
}
\description{
Makes an autocorrealtion plot using ggplot2 - which looks nicer than the
core R plot functions.
}

