% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggRunmean.R
\name{ggRunmean}
\alias{ggRunmean}
\title{ggRunmean}
\usage{
ggRunmean(output, pars, window.size = 10, title = "")
}
\arguments{
\item{dat}{The vector of the paramater you want to see an autocorrelation plot for.}
}
\description{
Make a running mean plot for an MCMC parameter using ggplot2.
}
