% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ggTrace.R
\name{ggTrace}
\alias{ggTrace}
\title{ggTrace}
\usage{
ggTrace(pars)
}
\arguments{
\item{pars}{The paramater you want to see an autocorrelation plot for.}
}
\description{
Make a trace plot for an MCMC parameter using ggplot2.
}

