% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funtions.R
\name{Litest}
\alias{Litest}
\title{calculate Div}
\usage{
Litest(microclim, SimTrait, ObsV, ObsF, InY)
}
\arguments{
\item{microclim}{microclim}

\item{SimTrait}{SimTrait}

\item{ObsV}{ObsV}

\item{ObsF}{ObsF}

\item{InY}{InY}

\item{param}{parameters

calculate Litest}
}
\value{
Div

Litest
}
\description{
calculate Div
}
