% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTrans.R
\name{makeTrans}
\alias{makeTrans}
\title{makeTrans}
\usage{
makeTrans(..., alpha = 0.5)
}
\arguments{
\item{alpha}{The alpha required.}
}
\description{
Some function from stack exchange that makes a colour transparent according to a given alpha.
}
