% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funtions.R
\name{modelEXP}
\alias{modelEXP}
\title{Regressiong Exp}
\usage{
modelEXP(dt, colx, coly, SSalpha, SSbeta, SSC)
}
\arguments{
\item{dt}{datas}

\item{colx}{X:Age}

\item{coly}{Y:MRW or maxLA}

\item{SSalpha}{alpha}

\item{SSbeta}{beta}

\item{SSC}{C value}
}
\value{
EXPmodel
}
\description{
Regressiong Exp
}
