% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplyNodes.R
\name{multiplyNodes}
\alias{multiplyNodes}
\title{multiplyNodes
Works out the cumulative effect of linear scalars on branches per iteration}
\usage{
multiplyNodes(scales, name, tree, Node_effects)
}
\arguments{
\item{scales}{A vector of scalars for a node}

\item{name}{The name of the node}

\item{tree}{The time tree}

\item{Node_effects}{A list, one element per node, to fill with the cumulative scalars}
}
\description{
multiplyNodes
Works out the cumulative effect of linear scalars on branches per iteration
}
\keyword{internal}
