% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funtions.R
\name{NewReplacesOld}
\alias{NewReplacesOld}
\title{replace parameters data}
\usage{
NewReplacesOld(DataNew, DataOld, ons = c("Parameter"))
}
\arguments{
\item{ons}{by on colume}

\item{DataA}{DataA}

\item{DataB}{DataB}
}
\value{
Pdata.frame
}
\description{
replace parameters data
}
