% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramLh.R
\name{paramLh}
\alias{paramLh}
\title{paramLh}
\usage{
paramLh(logfile, pars, cols = 2, thinning = 1, burnin = 0)
}
\arguments{
\item{logfile}{The name of the logfile of the BayesTraits run}

\item{pars}{A vector of the names of the parameters to be plotted against likelihood.}

\item{cols}{The number of columns of the resultant plot (if more than one plot)}

\item{thinning}{Thinning parameter for the posterior - defaults to 1 (all samples). 2 uses every second sample, 3 every third and so on.}

\item{burnin}{The number of generations to remove from the start of the chain as burnin.}
}
\description{
Plots a parameter against the likelihood for each iteration.
}
\examples{
paramsLh("cool-data.log", "Lambda")
paramsLh("cool-data.log", getParams("cool-data.log"))
params <- getParams("cool-data.log")
paramLh("cool-data.log", params[c(2:4)])
}
\keyword{log-likelihood}
\keyword{parameters}
