% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPhylo.R
\name{plotPhylo}
\alias{plotPhylo}
\title{plotPhylo}
\usage{
plotPhylo(tree, tips = FALSE, nodes = NULL, nodecex = NULL,
  scale = TRUE, ...)
}
\arguments{
\item{tree}{An object of class phylo}

\item{tips}{Logical - show tip labels or not?}

\item{nodes}{TRUE shows all node labels, FALSE supresses node labels (default), or a vecotr of which nodes to highlight.}

\item{nodecex}{Scaling factor for node labels.}

\item{scale}{Logical - show the scale bar, or not?}

\item{...}{Generic plot arguments (edge.width, edge.cols etc.)}
}
\description{
A silly little function that makes plotting a tree without tip labels and with
an axis easier and quicker.
}
