% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/oneParamPlot.R
\name{plotPosterior}
\alias{plotPosterior}
\title{plotPosterior}
\usage{
plotPosterior(logfile, params)
}
\arguments{
\item{logfile}{The name of the trait data file on which BayesTraits was run.}

\item{params}{A vector containing the name of the parameter(s) that are to be plotted. Can be
subsetted from the output of getParams().}
}
\description{
A function to plot one or more posterior distributions from the logfile of a BayesTraits
analysis.
}
\examples{
plotPosterior(cool-data.txt, c("Lh", "Alpha 1"))
plotPosterior(cool-data.txt, params[c(1:2)])
}
\keyword{distribution}
\keyword{histogram}
\keyword{plot}
\keyword{posterior}

