% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPosts.R
\name{plotPosts}
\alias{plotPosts}
\title{plotPosts}
\usage{
plotPosts(logfile, pars, fill = "dodgerblue", cols = 2, thinning = 1,
  burnin = 0)
}
\arguments{
\item{logfile}{The name of the trait data file on which BayesTraits was run.}

\item{fill}{The colour for the bars of the histogram. If "count" histogram will be shaded according to count.}

\item{cols}{The number of columns to plot multiple plots into.}

\item{thinning}{Thinning parameter for the posterior - defaults to 1 (all samples). 2 uses every second sample, 3 every third and so on.}

\item{burnin}{The number of generations to remove from the start of the chain as burnin.}

\item{params}{A vector containing the name of the parameter(s) that are to be plotted. Can be
subsetted from the output of getParams().}
}
\description{
A function to plot one or more posterior distributions from the logfile of a BayesTraits
analysis.
}
\examples{
plotPosterior(cool-data.txt, c("Lh", "Alpha 1"))
plotPosterior(cool-data.txt, params[c(1:2)])
}
\keyword{distribution}
\keyword{histogram}
\keyword{plot}
\keyword{posterior}
