% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotShifts.R
\name{rateShifts}
\alias{rateShifts}
\title{rateShifts
Generates the edge colours to colour edges by total rate.}
\usage{
rateShifts(PP, threshold, gradientcols, colour)
}
\description{
rateShifts
Generates the edge colours to colour edges by total rate.
}
\keyword{internal}
