% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunOfTrendAge.R
\name{RegData}
\alias{RegData}
\title{Regression funtion for Trend line outputs GAM & EXP}
\usage{
RegData(dt, Nage, Nrw, Nla, rwA, rwB, rwC, rwK, laA, laB, laC, laK)
}
\arguments{
\item{dt}{data}

\item{Nage}{colname of age}

\item{Nrw}{colname of ring width}

\item{Nla}{colname of La}

\item{rwA}{rwA}

\item{rwB}{rwB}

\item{rwC}{rwC}

\item{rwK}{rwK}

\item{laA}{laA}

\item{laB}{laB}

\item{laC}{laC}

\item{laK}{laK}
}
\value{
Trend line data frame
}
\description{
Regression funtion for Trend line outputs GAM & EXP
}
