% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funtions.R
\name{RegPlot}
\alias{RegPlot}
\title{Regression funtion for Trend line outputs GAM & EXP}
\usage{
RegPlot(RegData = dt, Mparam = dataIn$ManualParam, Oparam = dataIn$O_param)
}
\arguments{
\item{RegData}{data}

\item{Mparam}{colname of age}

\item{Oparam}{colname of ring width}
}
\value{
Trend line data frame
}
\description{
Regression funtion for Trend line outputs GAM & EXP
}
