% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rjParamCount.R
\name{rjParamCount}
\alias{rjParamCount}
\title{rjParamCount}
\usage{
rjParamCount(logs, burnin = 0, thinning = 1)
}
\arguments{
\item{logs}{A vector of one or more rj output logfile names.}

\item{burnin}{If burnin is needed for the output, set it here.}

\item{thinning}{If the psoterior needs thinning, set it here.}

\item{order}{Order the output. Logical - defaults to FALSE.}
}
\description{
Counts the mean number of parameters in a vecotr of one or more RJ models. Simpler than the full localscalarPP since just the parameter number is considered.
}

