% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjpp.R
\name{rjpp}
\alias{rjpp}
\title{rjpp}
\usage{
rjpp(rjlog, rjtrees, tree, burnin = 0, thinning = 1, meanbranches = TRUE,
  ratestable = TRUE)
}
\arguments{
\item{rjlog}{The RJ output of the run - typically suffixed with .VarRates.txt}

\item{tree}{The time tree the analysis was run on as an object of class "phylo", or the filename of the timetree.}

\item{burnin}{The burnin (if required) for the mcmc (generally worked out from the other logfile)}

\item{thinning}{Thinning parameter for the MCMC output - again, worked out from the raw MCMC output logfile.}

\item{meanbranches}{If true, calculates mean, median and mode branch lengths and returns mean tree.}

\item{ratestable}{}
}
\description{
A function that takes the output of a kappa, lambda, delta, VRates etc. RJ bayesTraits run and runs post-processing on it.
}
