% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_BTR.R
\name{sim_btr}
\alias{sim_btr}
\title{Sim BTR}
\usage{
sim_btr(param, clims = NA, Tage = NA, ObsF = NA, ObsV = NA, Obsline = NA)
}
\arguments{
\item{param}{parameters}

\item{clims}{climate}

\item{Tage}{trend of age}

\item{ObsF}{Observe Fiber}

\item{ObsV}{Observe vessel}

\item{Obsline}{取代ObsF V}

\item{syear}{start year}

\item{eyear}{end year}
}
\description{
Sim BTR
}
