% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_Cells.R
\name{Sim_Cells}
\alias{Sim_Cells}
\title{Simulate Fiber & Vessel cells}
\usage{
Sim_Cells(ObsV = NULL, ObsF = NULL, param, ...)
}
\arguments{
\item{ObsV}{Vessel anatomical trait measured by ROXAS}

\item{ObsF}{Fiber anatomical trait measured by ROXAS}

\item{param}{BTRmodels parameters data}
}
\description{
Simulate Fiber & Vessel cells
}
