% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_gRs.R
\name{Sim_gRs}
\alias{Sim_gRs}
\title{Sim gRs
use summary gR&gE and Standardization filtered RW sim threshold of clims}
\usage{
Sim_gRs(param, Tage, clims)
}
\arguments{
\item{param}{parameters}

\item{Tage}{trend of age}

\item{clims}{climate}
}
\description{
Sim gRs
use summary gR&gE and Standardization filtered RW sim threshold of clims
}
