% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funtions.R
\name{simTrend}
\alias{simTrend}
\title{Regressiong Exp}
\usage{
simTrend(Tdt, colx = "age", coly, method, O_param)
}
\arguments{
\item{colx}{X:Age}

\item{coly}{Y:MRW or maxLA}

\item{method}{methods}

\item{O_param}{Old parameters}

\item{dt}{datas}
}
\value{
EXPmodel
}
\description{
Regressiong Exp
}
