% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseModelString.R
\name{summariseModelString}
\alias{summariseModelString}
\title{summariseModelString}
\usage{
summariseModelString(logfile, thinning = 1, burnin = 0)
}
\arguments{
\item{logfile}{The name of the logfile resulting from an RJ BayesTraits model.}

\item{thinning}{Thinning parameter for the posterior - defaults to 1 (all samples). 2 uses every second sample, 3 every third and so on.}

\item{burnin}{The number of generations to remove from the start of the chain as burnin.}
}
\description{
Standardises the model string output from an RJ analysis, and returns a list of two dataframes.
One of these gives the frequency of each unique model and the parameter composition of it, and 
the other details the proportion of times each pair of parameters are locked to the same value.
This currently doesn't care whether the parameters are fixed to be the same, or whether they are
both zero - just that they are both the same.
}
