% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotShifts.R
\name{transShifts}
\alias{transShifts}
\title{transShifts
Generates the node labels, edge.colours and transparencies to plot the
location of transformations in a posterior.}
\usage{
transShifts(PP, threshold, cl, tree, transparency, relativetrans, nodescaling,
  colour, nodecex)
}
\description{
transShifts
Generates the node labels, edge.colours and transparencies to plot the
location of transformations in a posterior.
}
\keyword{internal}
