% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/viewLogfile.R
\name{viewLogfile}
\alias{viewLogfile}
\title{viewLogfile}
\usage{
viewLogfile(logfile, n = "max")
}
\arguments{
\item{logfile}{The name of the logfile.}

\item{n}{}
}
\description{
A simple function that shows the logfile.
}

