% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeroTest.R
\name{zeroTest}
\alias{zeroTest}
\title{zeroTest}
\usage{
zeroTest(logfile, pars, cols = 2, plot = TRUE, value = 0, thinning = 1,
  burnin = 0)
}
\arguments{
\item{logfile}{The name of the BayesTraits output logfile.}

\item{pars}{A vector of parameters to plot/summarise}

\item{cols}{The number of columns to plot into (if number of pars >1)}

\item{plot}{If true, a histogram coloured according to value is plotted. If false, a table summarising the posterior with respect to value is printed to screen.}

\item{value}{The value against which to test (defaults to zero)}

\item{thinning}{Thinning parameter for the posterior - defaults to 1 (all samples). 2 uses every second sample, 3 every third and so on.}

\item{burnin}{The number of generations to remove from the start of the chain as burnin.}
}
\description{
Takes a parameter and plots the posterior as a histogram with values >0 one
colour and values <0 another. If plot is turned off a table is returned
instead, which has, in decimal form, the fraction of values <0 and the 
fraction >0.
}
\examples{
zeroTest("cool-data.log", "Alpha")
zeroTest("cool-data.log", c("Alpha", "Beta.2"))
zeroTest("cool-data.log", "Lambda", value = 0.5)
zeroTest("cool-data.log", "Lambda", value = 0.5, plot = FALSE)
}
\keyword{posterior}
\keyword{significance}
\keyword{test}
\keyword{testing}
