% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compPosts.R
\name{compPosts}
\alias{compPosts}
\title{compPosts}
\usage{
compPosts(logs, pars, thinning = 1, burnin = 0, alpha = 0.5,
  fontsize = 3)
}
\arguments{
\item{logs}{The name of the trait data file on which BayesTraits was run, or a vector of >1 names if comparing between >1 logs.}

\item{pars}{A vector containing the names of two parameters to be compared. Must be a single parameter if comparing between two logs. Can 
be subsetted from the output of getParams().}

\item{thinning}{Thinning parameter for the posterior - defaults to 1 (all samples). 2 uses every second sample, 3 every third and so on.}

\item{burnin}{The number of generations to remove from the start of the chain as burnin.}
}
\description{
Compares histograms of one or more parameters from the same output file, or one parameter from one or more output files.
}
\examples{
plotPosterior(cool-data.txt, c("Lh", "Alpha 1"))
plotPosterior(cool-data.txt, params[c(1:2)])
}
\keyword{compare}
\keyword{distribution}
\keyword{histogram}
\keyword{plot}
\keyword{posterior}
