% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStones.R
\name{getStones}
\alias{getStones}
\title{getStones}
\usage{
getStones(logs, order = TRUE, paramcount = FALSE)
}
\arguments{
\item{logs}{A vector of one or more log files from stepping stones analysis.}
}
\description{
Get the marginal likelihoods from a single, or vector of, stepping stones log files.
}
