% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggDens.R
\name{ggDens}
\alias{ggDens}
\title{ggDens}
\usage{
ggDens(output, pars, title = "")
}
\arguments{
\item{dat}{The vector of paramater you want to see an autocorrelation plot for.}
}
\description{
Make a density plot for an MCMC parameter using ggplot2.
}
