% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggTrace.R
\name{ggTrace}
\alias{ggTrace}
\title{ggTrace}
\usage{
ggTrace(output, pars, title = "")
}
\arguments{
\item{output}{BayesTraits MCMC output - output from btmcmc function.}

\item{pars}{The paramater you want to see an autocorrelation plot for.}
}
\description{
Make a trace plot for an MCMC parameter using ggplot2.
}
