% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramDiagnostic.R
\name{paramDiagnostic}
\alias{paramDiagnostic}
\title{paramDiagnostic}
\usage{
paramDiagnostic(logfile, pars, type = NULL, cols = 2, thinning = 1,
  burnin = 0, title = NULL)
}
\arguments{
\item{logfile}{The name of the logfile of the BayesTraits run}

\item{pars}{A vector of the names of the parameters to generate diagnostic plots for.}

\item{type}{The type of diagnostic plot, if plotting across many parameters. Either "autocor", "dens", "trace" or "runmean"}

\item{cols}{The number of columns for multiple plotting.}

\item{thinning}{Thinning parameter for the posterior - defaults to 1 (all samples). 2 uses every second sample, 3 every third and so on.}

\item{burnin}{The number of generations to remove from the start of the chain as burnin.}
}
\description{
Plots 4 MCMC diagnostics for a parameter
}
\examples{
paramDiagnostic("cool-data.log", "Lh")
params <- getParams("cool-data.log")
paramDiagnostic("cool-data.log", params, type = "trace", cols = 3)
}
\keyword{autocorrelation}
\keyword{density}
\keyword{diagnostic}
\keyword{mcmc}
\keyword{mean}
\keyword{running}
\keyword{trace}
