% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewLogfile.R
\name{viewLogfile}
\alias{viewLogfile}
\title{viewLogfile}
\usage{
viewLogfile(logfile, n = "max", thinning = 1, burnin = 0)
}
\arguments{
\item{logfile}{The name of the logfile.}

\item{n}{The number of rows of the logfile to view (defaults to the whole file)}

\item{thinning}{Thinning parameter for the posterior - defaults to 1 (all samples). 2 uses every second sample, 3 every third and so on.}
}
\description{
A simple function that shows the logfile.
}
